Adding support for the retrieval of the object-ring and the object-stored-ring.
The object-ring is the ring containing all objects that are linked.  It will fetch the needed objects from the database *before* storing the result.  This will adhere to the concept of *-slot-many and *-slot-one.
The object-stored-ring contains all the followable linked objects that have been stored in the quickstore.  It does not adhere to the *-one-slot and will return all the slots that fit the picture (instead of one for *-slot-one).